RotationMNIST/requirements.txt
backpack==0.1
clip==0.2.0
numpy==1.26.4
pandas==2.2.2
Pillow==9.5.0
Pillow==10.3.0
scikit_learn==1.3.2
scipy==1.13.0
six==1.16.0
tensorflow==2.16.1
timm==0.9.12
torch==1.12.0+cu116
torch_optimizer==0.3.0
torchvision==0.13.0+cu116
tqdm==4.65.0
